sound: soc: rhine: use correct plug detection values
In rhine stock kernel, there are two types of values for detection. There are values for noise dumping jack and non-noise dumping jack.
It would make sense to use values for non-dumping jack, the problem is that the values are not correct. Detection does not work and mbhc detects all the jacks as 4-pole ones. That results in not working microphone while headphones connected and phantom button presses.
Values for noise dumping jack work correctly, phone mic works and there are no phantom presses.